requirements.txt
# Telco Customer Churn - ML Project Dependencies
# Install: pip install -r requirements.txt
# Create virtual env: python -m venv venv && source venv/bin/activate (Unix) or venv\Scripts\activate (Windows)

pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
kaggle>=1.5.0
python-dotenv>=1.0.0
ydata-profiling>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
